corrections afer code review; new BackButton component
