+++ src/components/Blogtag/BlogTag.js
@@ -7,18 +7,16 @@ export default function BlogTag({tagName}) {
   const colorVar = tagColors.find(col => col.tag === tagName).color;
 
   return (
-    <>
-      <Button
-        style={{backgroundColor: colorVar}}
-        role="navigation"
-        aria-label={`Zeige alle Artikel und Videos zum Thema ${tagName}`}
-        onClick={() => {
-          navigate(`/search/${tagName}`.toLowerCase());
-        }}
-      >
-        {tagName}
-      </Button>
-    </>
+    <Button
+      style={{backgroundColor: colorVar}}
+      role="navigation"
+      aria-label={`Zeige alle Artikel und Videos zum Thema ${tagName}`}
+      onClick={() => {
+        navigate(`/search/${tagName}`.toLowerCase());
+      }}
+    >
+      {tagName}
+    </Button>
   );
 }
 
@@ -38,12 +36,7 @@ const Button = styled.button`
 
 export function BlogTagLine({tagName}) {
   const colorVar = tagColors.find(col => col.tag === tagName).color;
-
-  return (
-    <>
-      <Paragraph style={{backgroundColor: colorVar ? colorVar : 'black'}}>{tagName}</Paragraph>
-    </>
-  );
+  return <Paragraph style={{backgroundColor: colorVar ? colorVar : 'black'}}>{tagName}</Paragraph>;
 }
 
 const Paragraph = styled.div`
